Added api for uploading and downloading images to the database
Earlier there was no way for storing and reading
images, now images can be saved as base64 strings
on the database, in the Images collection. These
images are meant to be unique, user uploaded
assets. Images that are repetitive and not unique
to any user are more appropriate to server
publicly from the server itself.