Integrated fetching of userlinks data using username or session tokens
Earlier the server was setup to receive the
username and it would return the userlinks data,
now the session tokens api also works and can be
used to find the username and then the username
can be used to get userlinks data as normal. Most
of the changes made were actually made on the
frontend and not on the server side.