Modified the createToken script to search tokens using username
Earlier the findToken function was being called
from withing the createToken function, to search
for a pre existing token with a given username.
This search was always bound to fail since
findToken works using the hashed token and not
using the username. Therefore, now the createToken
function performs its own search using the passed
username.